+++ src/test/ExSln2/c99/gen/hal_LedArray.c
@@ -12,7 +12,7 @@ void hal_LedArray_ctor(hal_LedArray * self, hal_Led * * leds, uint8_t leds_lengt
     self->_leds_length = leds_length;
 }
 
-hal_Led * hal_LedArray_getLed(hal_LedArray * self, uint8_t index)
+hal_Led * hal_LedArray_get_led(hal_LedArray * self, uint8_t index)
 {
     return self->_leds[index];
 }